split up the massive AppNavigation.tsx
