+++ src/features/navigation/AppNavigation.module.css
@@ -16,354 +16,3 @@
 .group-list > li:not(:last-child) {
   margin-bottom: 28px;
 }
-
-.task-button {
-  font-family: var(--font-family);
-  font-size: 16px;
-  background-color: transparent;
-  border: none;
-  cursor: not-allowed;
-  display: grid;
-  width: 100%;
-  grid-template-columns: auto 1fr auto;
-  column-gap: 8px;
-  align-items: start;
-  text-align: left;
-
-  border-radius: var(--fds-border_radius-interactive);
-  padding: 2px;
-  margin: -2px;
-}
-
-.group-name {
-  min-height: 28px;
-  display: inline-block;
-  align-content: center;
-  line-height: 1.2;
-}
-
-.group-name-active {
-  text-decoration-line: underline;
-  text-decoration-skip-ink: none;
-  text-underline-offset: 3px;
-  text-decoration-thickness: 1px;
-}
-
-.task-symbol-locked {
-  color: var(--fds-colors-grey-800);
-  border-color: var(--fds-colors-grey-600);
-  background-color: var(--fds-colors-grey-100);
-}
-
-.task-symbol-active {
-  color: var(--fds-colors-grey-100);
-  border-color: var(--fds-colors-blue-900);
-  background-color: var(--fds-colors-blue-900);
-}
-
-.group-button {
-  font-family: var(--font-family);
-  font-size: 16px;
-  font-weight: normal;
-  color: var(--fds-colors-grey-800);
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-  display: grid;
-  grid-template-columns: auto auto auto;
-  column-gap: 8px;
-  align-items: start;
-  text-align: left;
-
-  border-radius: var(--fds-border_radius-interactive);
-  padding: 2px 6px 2px 2px;
-  margin: -2px;
-
-  position: relative;
-}
-
-.group-button:disabled {
-  opacity: 0.7;
-  cursor: not-allowed;
-}
-
-.group-button:hover:not(:disabled) {
-  text-decoration: underline;
-  text-decoration-skip-ink: none;
-  text-underline-offset: 2px;
-  text-decoration-thickness: 1px;
-}
-
-.group-button-single {
-  color: var(--fds-colors-blue-700);
-  grid-template-columns: auto auto;
-}
-
-.group-chevron {
-  margin-top: 6px;
-  margin-right: -4px; /* The icon has some internal padding, this makes the edge line up better */
-  color: var(--fds-colors-grey-700);
-  font-size: 16px;
-}
-
-.group-chevron-open {
-  transform: rotate(180deg);
-}
-
-.group-symbol {
-  width: 28px;
-  height: 28px;
-  box-sizing: border-box;
-  font-size: 20px;
-  border-width: 2px;
-  border-style: solid;
-  border-radius: var(--fds-border_radius-interactive);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.group-symbol-single {
-  font-size: 18px;
-  border-radius: 50%;
-  width: 24px;
-  height: 24px;
-  margin: 2px;
-}
-
-.group-symbol-error {
-  color: var(--fds-colors-red-700);
-  border-color: var(--fds-colors-red-500);
-  background-color: var(--fds-colors-red-100);
-}
-
-.group-symbol-complete {
-  color: var(--fds-colors-grey-100);
-  border-color: var(--fds-colors-blue-700);
-  background-color: var(--fds-colors-blue-700);
-}
-
-.group-symbol-active {
-  color: var(--fds-colors-grey-100);
-  border-color: var(--fds-colors-blue-900);
-  background-color: var(--fds-colors-blue-900);
-}
-
-.group-symbol-active.group-symbol-single:not(.group-symbol-info) {
-  outline: 2px solid var(--fds-colors-grey-100);
-  outline-offset: -4px;
-}
-
-.group-symbol-default {
-  color: var(--fds-colors-blue-900);
-  border-color: var(--fds-colors-blue-400);
-  background-color: var(--fds-colors-grey-100);
-}
-
-.page-list {
-  list-style-type: none;
-  padding: 0;
-  margin: 0;
-}
-
-.page-list-item {
-  position: relative;
-}
-
-.page-list-item:first-child {
-  margin-top: 15px;
-}
-
-.page-list-item:not(:last-child) {
-  margin-bottom: 20px;
-}
-
-.group-button-open::before {
-  content: '';
-  position: absolute;
-  box-sizing: border-box;
-  width: 0;
-  border-left: 2px solid var(--fds-colors-grey-700);
-  height: calc(100% - 21px);
-  top: 33px;
-  left: 15px;
-}
-
-.page-list-item:not(:last-child)::after {
-  content: '';
-  position: absolute;
-  box-sizing: border-box;
-  width: 0;
-  border-left: 2px solid var(--fds-colors-grey-700);
-  height: calc(100% - 8px);
-  top: 25px;
-  left: 13px;
-}
-
-.page-button {
-  font-family: var(--font-family);
-  font-size: 16px;
-  font-weight: normal;
-  color: var(--fds-colors-blue-700);
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-  display: grid;
-  grid-template-columns: auto auto;
-  column-gap: 12px;
-  align-items: start;
-  text-align: left;
-
-  border-radius: var(--fds-border_radius-interactive);
-  padding: 2px 6px;
-  margin: -2px;
-}
-
-.page-button:disabled {
-  opacity: 0.7;
-  cursor: not-allowed;
-}
-
-.page-button:hover:not(:disabled) {
-  text-decoration-line: underline;
-  text-decoration-skip-ink: none;
-  text-underline-offset: 2px;
-  text-decoration-thickness: 1px;
-}
-
-.page-symbol {
-  width: 20px;
-  height: 20px;
-  box-sizing: border-box;
-  border-width: 2px;
-  border-style: solid;
-  border-radius: 50%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.page-symbol-error {
-  font-size: 14px;
-  color: var(--fds-colors-red-700);
-  border-color: var(--fds-colors-red-500);
-  background-color: var(--fds-colors-red-100);
-}
-
-.page-symbol-complete {
-  font-size: 14px;
-  color: white;
-  border-color: var(--fds-colors-blue-700);
-  background-color: var(--fds-colors-blue-700);
-}
-
-.page-symbol-default {
-  border-color: var(--fds-colors-blue-400);
-  background-color: var(--fds-colors-grey-100);
-}
-
-.page-symbol-active {
-  border-color: var(--fds-colors-blue-900);
-  background-color: var(--fds-colors-blue-900);
-
-  outline: 2px solid var(--fds-colors-grey-100);
-  outline-offset: -4px;
-}
-
-.page-name {
-  margin-top: 1px;
-  line-height: 1.2;
-}
-.page-name-active {
-  text-decoration-line: underline;
-  text-decoration-skip-ink: none;
-  text-underline-offset: 3px;
-  text-decoration-thickness: 1px;
-}
-
-.subform-container {
-  margin-top: 16px;
-  margin-left: 32px;
-}
-
-.subform-expand-button {
-  font-family: var(--font-family);
-  font-size: 16px;
-  font-weight: normal;
-  color: var(--fds-colors-grey-700);
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-  display: grid;
-  grid-template-columns: auto auto;
-  column-gap: 8px;
-  align-items: start;
-  text-align: left;
-
-  border-radius: var(--fds-border_radius-interactive);
-  padding: 2px 6px;
-  margin: -2px;
-}
-
-.subform-expand-button:hover {
-  text-decoration: underline;
-  text-decoration-skip-ink: none;
-  text-underline-offset: 2px;
-  text-decoration-thickness: 1px;
-}
-
-.subform-expand-chevron {
-  margin-top: 1px;
-  margin-right: -4px; /* The icon has some internal padding, this makes the edge line up better */
-  color: var(--fds-colors-grey-700);
-  font-size: 16px;
-}
-
-.subform-expand-chevron-open {
-  transform: rotate(180deg);
-}
-
-.subform-group-name {
-  line-height: 1.2;
-}
-
-.subform-list {
-  list-style-type: none;
-  padding: 0;
-  margin: 12px 0 0 0;
-  display: flex;
-  flex-direction: column;
-  row-gap: 8px;
-}
-
-.subform-link {
-  display: block;
-  font-family: var(--font-family);
-  font-size: 14px;
-  font-weight: normal;
-  color: var(--fds-colors-blue-700);
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-  text-align: left;
-
-  border-radius: var(--fds-border_radius-interactive);
-  padding: 2px 6px;
-  margin: -2px;
-}
-
-.subform-link:disabled {
-  opacity: 0.7;
-  cursor: not-allowed;
-}
-
-.subform-link:hover:not(:disabled) {
-  text-decoration-line: underline;
-  text-decoration-skip-ink: none;
-  text-underline-offset: 2px;
-  text-decoration-thickness: 1px;
-}
-
-.subform-link-name {
-  line-height: 1.2;
-}
