+++ src/features/navigation/components/Page.module.css
@@ -1,0 +1,104 @@
+.page-list-item {
+  position: relative;
+}
+
+.page-list-item:first-child {
+  margin-top: 15px;
+}
+
+.page-list-item:not(:last-child) {
+  margin-bottom: 20px;
+}
+
+.page-list-item:not(:last-child)::after {
+  content: '';
+  position: absolute;
+  box-sizing: border-box;
+  width: 0;
+  border-left: 2px solid var(--fds-colors-grey-700);
+  height: calc(100% - 8px);
+  top: 25px;
+  left: 13px;
+}
+
+.page-button {
+  font-family: var(--font-family);
+  font-size: 16px;
+  font-weight: normal;
+  color: var(--fds-colors-blue-700);
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  display: grid;
+  grid-template-columns: auto auto;
+  column-gap: 12px;
+  align-items: start;
+  text-align: left;
+
+  border-radius: var(--fds-border_radius-interactive);
+  padding: 2px 6px;
+  margin: -2px;
+}
+
+.page-button:disabled {
+  opacity: 0.7;
+  cursor: not-allowed;
+}
+
+.page-button:hover:not(:disabled) {
+  text-decoration-line: underline;
+  text-decoration-skip-ink: none;
+  text-underline-offset: 2px;
+  text-decoration-thickness: 1px;
+}
+
+.page-symbol {
+  width: 20px;
+  height: 20px;
+  box-sizing: border-box;
+  border-width: 2px;
+  border-style: solid;
+  border-radius: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.page-symbol-error {
+  font-size: 14px;
+  color: var(--fds-colors-red-700);
+  border-color: var(--fds-colors-red-500);
+  background-color: var(--fds-colors-red-100);
+}
+
+.page-symbol-complete {
+  font-size: 14px;
+  color: white;
+  border-color: var(--fds-colors-blue-700);
+  background-color: var(--fds-colors-blue-700);
+}
+
+.page-symbol-default {
+  border-color: var(--fds-colors-blue-400);
+  background-color: var(--fds-colors-grey-100);
+}
+
+.page-symbol-active {
+  border-color: var(--fds-colors-blue-900);
+  background-color: var(--fds-colors-blue-900);
+
+  outline: 2px solid var(--fds-colors-grey-100);
+  outline-offset: -4px;
+}
+
+.page-name {
+  margin-top: 1px;
+  line-height: 1.2;
+}
+
+.page-name-active {
+  text-decoration-line: underline;
+  text-decoration-skip-ink: none;
+  text-underline-offset: 3px;
+  text-decoration-thickness: 1px;
+}
