+++ src/features/navigation/components/PageGroup.module.css
@@ -1,0 +1,129 @@
+.group-name {
+  min-height: 28px;
+  display: inline-block;
+  align-content: center;
+  line-height: 1.2;
+}
+
+.group-name-active {
+  text-decoration-line: underline;
+  text-decoration-skip-ink: none;
+  text-underline-offset: 3px;
+  text-decoration-thickness: 1px;
+}
+
+.group-button {
+  font-family: var(--font-family);
+  font-size: 16px;
+  font-weight: normal;
+  color: var(--fds-colors-grey-800);
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  display: grid;
+  grid-template-columns: auto auto auto;
+  column-gap: 8px;
+  align-items: start;
+  text-align: left;
+
+  border-radius: var(--fds-border_radius-interactive);
+  padding: 2px 6px 2px 2px;
+  margin: -2px;
+
+  position: relative;
+}
+
+.group-button:disabled {
+  opacity: 0.7;
+  cursor: not-allowed;
+}
+
+.group-button:hover:not(:disabled) {
+  text-decoration: underline;
+  text-decoration-skip-ink: none;
+  text-underline-offset: 2px;
+  text-decoration-thickness: 1px;
+}
+
+.group-button-single {
+  color: var(--fds-colors-blue-700);
+  grid-template-columns: auto auto;
+}
+
+.group-chevron {
+  margin-top: 6px;
+  margin-right: -4px; /* The icon has some internal padding, this makes the edge line up better */
+  color: var(--fds-colors-grey-700);
+  font-size: 16px;
+}
+
+.group-chevron-open {
+  transform: rotate(180deg);
+}
+
+.group-symbol {
+  width: 28px;
+  height: 28px;
+  box-sizing: border-box;
+  font-size: 20px;
+  border-width: 2px;
+  border-style: solid;
+  border-radius: var(--fds-border_radius-interactive);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.group-symbol-single {
+  font-size: 18px;
+  border-radius: 50%;
+  width: 24px;
+  height: 24px;
+  margin: 2px;
+}
+
+.group-symbol-error {
+  color: var(--fds-colors-red-700);
+  border-color: var(--fds-colors-red-500);
+  background-color: var(--fds-colors-red-100);
+}
+
+.group-symbol-complete {
+  color: var(--fds-colors-grey-100);
+  border-color: var(--fds-colors-blue-700);
+  background-color: var(--fds-colors-blue-700);
+}
+
+.group-symbol-active {
+  color: var(--fds-colors-grey-100);
+  border-color: var(--fds-colors-blue-900);
+  background-color: var(--fds-colors-blue-900);
+}
+
+.group-symbol-active.group-symbol-single:not(.group-symbol-info) {
+  outline: 2px solid var(--fds-colors-grey-100);
+  outline-offset: -4px;
+}
+
+.group-symbol-default {
+  color: var(--fds-colors-blue-900);
+  border-color: var(--fds-colors-blue-400);
+  background-color: var(--fds-colors-grey-100);
+}
+
+.page-list {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+}
+
+.group-button-open::before {
+  content: '';
+  position: absolute;
+  box-sizing: border-box;
+  width: 0;
+  border-left: 2px solid var(--fds-colors-grey-700);
+  height: calc(100% - 21px);
+  top: 33px;
+  left: 15px;
+}
