+++ src/features/navigation/components/SubformsForPage.module.css
@@ -1,0 +1,86 @@
+.subform-container {
+  margin-top: 16px;
+  margin-left: 32px;
+}
+
+.subform-expand-button {
+  font-family: var(--font-family);
+  font-size: 16px;
+  font-weight: normal;
+  color: var(--fds-colors-grey-700);
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  display: grid;
+  grid-template-columns: auto auto;
+  column-gap: 8px;
+  align-items: start;
+  text-align: left;
+
+  border-radius: var(--fds-border_radius-interactive);
+  padding: 2px 6px;
+  margin: -2px;
+}
+
+.subform-expand-button:hover {
+  text-decoration: underline;
+  text-decoration-skip-ink: none;
+  text-underline-offset: 2px;
+  text-decoration-thickness: 1px;
+}
+
+.subform-expand-chevron {
+  margin-top: 1px;
+  margin-right: -4px; /* The icon has some internal padding, this makes the edge line up better */
+  color: var(--fds-colors-grey-700);
+  font-size: 16px;
+}
+
+.subform-expand-chevron-open {
+  transform: rotate(180deg);
+}
+
+.subform-group-name {
+  line-height: 1.2;
+}
+
+.subform-list {
+  list-style-type: none;
+  padding: 0;
+  margin: 12px 0 0 0;
+  display: flex;
+  flex-direction: column;
+  row-gap: 8px;
+}
+
+.subform-link {
+  display: block;
+  font-family: var(--font-family);
+  font-size: 14px;
+  font-weight: normal;
+  color: var(--fds-colors-blue-700);
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+  text-align: left;
+
+  border-radius: var(--fds-border_radius-interactive);
+  padding: 2px 6px;
+  margin: -2px;
+}
+
+.subform-link:disabled {
+  opacity: 0.7;
+  cursor: not-allowed;
+}
+
+.subform-link:hover:not(:disabled) {
+  text-decoration-line: underline;
+  text-decoration-skip-ink: none;
+  text-underline-offset: 2px;
+  text-decoration-thickness: 1px;
+}
+
+.subform-link-name {
+  line-height: 1.2;
+}
